pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mimosa"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "jax",
    "equinox",
    "optimistix",
    "kernax-ml",
    "matplotlib",
]